document-automation/requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
gunicorn==23.0.0
python-multipart==0.0.9
jinja2==3.1.4

# Task Queue
celery==5.4.0
redis==5.0.8

# Database
sqlalchemy==2.0.35
alembic==1.13.3
psycopg2-binary==2.9.9

# Google Cloud APIs
google-cloud-vision==3.7.4
google-generativeai==0.8.2

# File Processing
pillow==10.4.0
pdf2image==1.17.0
pypdf2==3.0.1
python-magic==0.4.27

# OCR (Fallback Local)
pytesseract==0.3.13

# Data Processing
pandas==2.2.3
openpyxl==3.1.5

# Utilities
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.2
aiofiles==24.1.0
markdown==3.7

# Logging & Monitoring
python-json-logger==2.0.7

# RAG System Dependencies
qdrant-client==1.7.0
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
sentence-transformers==2.7.0
huggingface-hub==0.20.3
openai==1.12.0
tiktoken==0.5.2
numpy==1.24.3
scikit-learn==1.3.2

